update & create animal
